Allow extra chests in exchange for unused dispensers.
This patch allows to automatically extend the upper limit for chests and trapped chests, adding one extra chest for every three dispensers not used (but allowed) on the craft.

No other block limit is changed (including the craft size) and it won't affect any of the current crafts (but allows to manually remove dispensers and replace them with chests).

Additionally, the detection process now gives more information when a fly block limit is not met (not only percentages but also block counts and craft size). It also reports all the violations found (not only the first one). Examples:

    Not enough flyblock: redstone block -> 9 < 10 (1,00% of 1000)

or

    Too much flyblock: chest -> 31 > 29 (3,00% of 999) + 1 extra (in exchange for 3 dispensers)
    Not enough flyblock: redstone block -> 8 < 9 (1,00% of 999)

This feature is controlled by the config entry `ExchangeDispensersForChests` which is `false` by default (for backward compatibility).